requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pyaml==20.4.0
PyYAML==5.3.1
rauth==0.7.3
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
yahoo-oauth==1.0
yarl==1.5.1
